Add contenteditable tagging implementation
There are a whole bunch of small unrelated changes here I can't be
bothered to isolate, so deal with this messy commit.

Basically this displays the text and tags as a header and a list of
items beneath it, and sets them as contenteditable for a simple UI.
Adding new text or tags can be done by typing and pressing Enter or
Tab (or just clicking outside of the element (maybe I should change
this?)).