Move Meteor files into app/ directory
Meteor is such a douchebag that there's no way for you to tell it to
ignore certain files or directories in the root dir. So if you want to
use Grunt or Gulp for task automation, you're screwed, since Meteor will
try to read node_modules/ and go bonkers. Great job, Meteor devs.