Fix: avoid encoding URLs twice
Previously, iron-router would encode paths twice, so passing an already
encoded URL as a route argument would encode it twice. For example,
'http://' would render as 'http%253A%252F%252F' instead of
'http%3A%2F%2F'. So I chose to avoid using the pathFor helper, which is
a bummer for maintaining the routes in the future.